fix(contracts): secure config_system, widen daily scores, fix deploy configs
- config_system: add OpenZeppelin AccessControlComponent with ADMIN_ROLE.
  Global setters (set_treasury, set_zstar_address, set_star_eligible)
  require ADMIN_ROLE. Per-settings setters (set_zone_pricing,
  set_zone_enabled, set_zone_theme) require owner or ADMIN_ROLE.
  dojo_init grants both DEFAULT_ADMIN_ROLE and ADMIN_ROLE to
  creator_address without changing init args.

- Daily score pipeline: DailyEntry.best_score u16→u32,
  DailyLeaderboard.value u32→u64. Ranking composite changed from
  stars*65536+score to (stars<<32)|score. Updated game_over,
  daily_challenge, and daily helper end-to-end.

- torii_sepolia.toml namespace fixed to zkube_v2_1_0.
  Zero-address tokens in dojo_mainnet.toml marked as deploy blockers.
  contracts/.tool-versions synced with root.

- Client: "Connect Wallet" → "Connect Account".

- Added test_config_auth and test_daily_scores regression tests.

+++ contracts/src/helpers/daily.cairo
@@ -18,7 +18,7 @@ pub const MAX_LEADERBOARD_SIZE: u32 = 250;
 /// - game_over::auto_submit_daily_result (inline during game over)
 /// - daily_challenge_system::submit_result (manual backup submission)
 pub fn update_daily_leaderboard(
-    ref world: WorldStorage, challenge_id: u32, player: ContractAddress, new_value: u32,
+    ref world: WorldStorage, challenge_id: u32, player: ContractAddress, new_value: u64,
 ) {
     // Walk the leaderboard from rank 1 to find insertion point
     let mut insert_rank: u32 = 0;
